+++ Snowdog_Menu/requirejs-config.js
@@ -1,6 +1,7 @@
 var config = {
     map: {
         "*": {
+            "megaMenu": "Snowdog_Menu/js/mega-menu.babel",
             "sideMenu": "Snowdog_Menu/js/side-menu.babel"
         }
     }